# backend/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Project-350 Travel Assistant API",
    "description": "Comprehensive API documentation for the AI-powered travel assistant platform",
    "version": "1.0.0",
    "contact": {
      "name": "Project-350 Team",
      "email": "support@project350.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Bearer token. Format: Bearer {token}"
    }
  },
  "paths": {
    "/api/trips": {
      "post": {
        "tags": ["Trips"],
        "summary": "Create a new trip",
        "description": "Create a new travel itinerary with AI-generated suggestions",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["destination", "duration", "budget"],
              "properties": {
                "destination": {
                  "type": "string",
                  "example": "Cox's Bazar, Bangladesh"
                },
                "duration": {
                  "type": "integer",
                  "example": 3,
                  "description": "Duration in days"
                },
                "budget": {
                  "type": "number",
                  "example": 500,
                  "description": "Budget in USD"
                },
                "preferences": {
                  "type": "array",
                  "items": {"type": "string"},
                  "example": ["beach", "adventure", "culture"]
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Trip created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {"type": "boolean"},
                "trip": {"$ref": "#/definitions/Trip"}
              }
            }
          },
          "400": {"description": "Bad Request - Invalid input"},
          "401": {"description": "Unauthorized"},
          "500": {"description": "Internal Server Error"}
        }
      }
    },
    "/api/trips/{id}": {
      "get": {
        "tags": ["Trips"],
        "summary": "Get trip by ID",
        "description": "Retrieve a specific trip with all details",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Trip ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Trip retrieved successfully",
            "schema": {"$ref": "#/definitions/Trip"}
          },
          "404": {"description": "Trip not found"},
          "401": {"description": "Unauthorized"}
        }
      },
      "delete": {
        "tags": ["Trips"],
        "summary": "Delete a trip",
        "description": "Delete a specific trip by ID",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Trip ID to delete"
          }
        ],
        "responses": {
          "200": {"description": "Trip deleted successfully"},
          "404": {"description": "Trip not found"},
          "401": {"description": "Unauthorized"},
          "500": {"description": "Internal Server Error"}
        }
      }
    },
    "/api/blogs": {
      "get": {
        "tags": ["Blogs"],
        "summary": "Get all blogs",
        "description": "Retrieve all travel blogs with pagination",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "description": "Page number for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "description": "Number of blogs per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Blogs retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "blogs": {
                  "type": "array",
                  "items": {"$ref": "#/definitions/Blog"}
                },
                "pagination": {"$ref": "#/definitions/Pagination"}
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Blogs"],
        "summary": "Create a blog",
        "description": "Create a new travel blog from trip experience",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["tripId"],
              "properties": {
                "tripId": {
                  "type": "string",
                  "example": "trip_123456"
                },
                "title": {
                  "type": "string",
                  "example": "Amazing Adventure in Cox's Bazar"
                },
                "content": {
                  "type": "string",
                  "example": "Custom content for the blog..."
                }
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "Blog created successfully"},
          "400": {"description": "Bad Request"},
          "401": {"description": "Unauthorized"},
          "500": {"description": "Internal Server Error"}
        }
      },
      "put": {
        "tags": ["Blogs"],
        "summary": "Update a blog",
        "description": "Update an existing blog",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "blogId": {"type": "string"},
                "title": {"type": "string"},
                "content": {"type": "string"}
              }
            }
          }
        ],
        "responses": {
          "200": {"description": "Blog updated successfully"},
          "400": {"description": "Bad Request"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Blog not found"}
        }
      }
    },
    "/api/blogs/{id}": {
      "delete": {
        "tags": ["Blogs"],
        "summary": "Delete a blog",
        "description": "Delete a specific blog by ID",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Blog ID"
          }
        ],
        "responses": {
          "200": {"description": "Blog deleted successfully"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Blog not found"}
        }
      }
    },
    "/api/blogs/{id}/comments": {
      "post": {
        "tags": ["Blogs"],
        "summary": "Add comment to blog",
        "description": "Add a comment to a specific blog",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Blog ID"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": ["comment"],
              "properties": {
                "comment": {
                  "type": "string",
                  "example": "Great travel experience!"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "Comment added successfully"},
          "400": {"description": "Bad Request"},
          "401": {"description": "Unauthorized"},
          "404": {"description": "Blog not found"}
        }
      }
    },
    "/api/blogs/{blogId}/comments/{commentId}/replies": {
      "post": {
        "tags": ["Blogs"],
        "summary": "Reply to comment",
        "description": "Add a reply to a specific comment",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": ["reply"],
              "properties": {
                "reply": {
                  "type": "string",
                  "example": "Thank you for sharing!"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "Reply added successfully"},
          "400": {"description": "Bad Request"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
    "/api/blogs/{id}/like": {
      "post": {
        "tags": ["Blogs"],
        "summary": "Like a blog",
        "description": "Like or unlike a blog post",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Blog ID"
          }
        ],
        "responses": {
          "200": {"description": "Like status updated"},
          "401": {"description": "Unauthorized"},
          "404": {"description": "Blog not found"}
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Authenticate user and return JWT token",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "password"],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "user@example.com"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "example": "securepassword123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "success": {"type": "boolean"},
                "token": {"type": "string"},
                "user": {"$ref": "#/definitions/User"}
              }
            }
          },
          "400": {"description": "Bad Request - Missing credentials"},
          "401": {"description": "Invalid credentials"},
          "500": {"description": "Internal Server Error"}
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User registration",
        "description": "Register a new user account",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "email", "password"],
              "properties": {
                "name": {
                  "type": "string",
                  "example": "John Doe"
                },
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "john@example.com"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "example": "securepassword123",
                  "minLength": 6
                }
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "User registered successfully"},
          "400": {"description": "Bad Request - Invalid input"},
          "409": {"description": "User already exists"},
          "500": {"description": "Internal Server Error"}
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User logout",
        "description": "Logout user and invalidate token",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Logout successful"}
        }
      }
    },
    "/api/auth/profile/update": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Update user profile",
        "description": "Update user profile information",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {"type": "string", "example": "John Doe"},
                "username": {"type": "string", "example": "johndoe"},
                "bio": {"type": "string", "example": "Travel enthusiast"},
                "location": {"type": "string", "example": "New York, USA"}
              }
            }
          }
        ],
        "responses": {
          "200": {"description": "Profile updated successfully"},
          "400": {"description": "Bad Request"},
          "401": {"description": "Unauthorized"},
          "404": {"description": "User not found"}
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh JWT token",
        "description": "Get a new access token using refresh token",
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "schema": {
              "type": "object",
              "properties": {
                "accessToken": {"type": "string"}
              }
            }
          },
          "401": {"description": "Invalid refresh token"}
        }
      }
    },
    "/api/auth/send-verification-otp": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Send email verification OTP",
        "description": "Send a verification OTP to user's email address",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email"],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "user@example.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {"type": "boolean"},
                "message": {"type": "string"},
                "messageId": {"type": "string"}
              }
            }
          },
          "400": {"description": "Bad Request - Email required or already verified"},
          "404": {"description": "User not found"},
          "500": {"description": "Internal Server Error"}
        }
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Verify email with OTP",
        "description": "Verify user's email address using the OTP code",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "otp"],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "user@example.com"
                },
                "otp": {
                  "type": "string",
                  "pattern": "^[0-9]{6}$",
                  "example": "123456",
                  "description": "6-digit verification code"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {"type": "boolean"},
                "message": {"type": "string"},
                "user": {"$ref": "#/definitions/User"}
              }
            }
          },
          "400": {"description": "Bad Request - Invalid or expired OTP"},
          "404": {"description": "User not found"},
          "500": {"description": "Internal Server Error"}
        }
      }
    },
    "/api/auth/resend-verification-otp": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Resend verification OTP",
        "description": "Resend verification OTP to user's email (with cooldown protection)",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email"],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "user@example.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP resent successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {"type": "boolean"},
                "message": {"type": "string"}
              }
            }
          },
          "400": {"description": "Bad Request - Email already verified or cooldown active"},
          "404": {"description": "User not found"},
          "500": {"description": "Internal Server Error"}
        }
      }
    },
    "/api/auth/otp-status": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Check OTP status",
        "description": "Check if user has a pending OTP and get status information",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          {
            "name": "purpose",
            "in": "query",
            "type": "string",
            "default": "registration",
            "enum": ["registration", "password_reset", "email_verification", "login_verification"],
            "description": "Purpose of the OTP"
          }
        ],
        "responses": {
          "200": {
            "description": "OTP status retrieved",
            "schema": {
              "type": "object",
              "properties": {
                "success": {"type": "boolean"},
                "hasPendingOTP": {"type": "boolean"},
                "timeLeft": {"type": "integer", "description": "Minutes remaining before expiry"},
                "attempts": {"type": "integer"},
                "maxAttempts": {"type": "integer"}
              }
            }
          },
          "400": {"description": "Bad Request - Email required"},
          "500": {"description": "Internal Server Error"}
        }
      }
    },
    "/api/groups": {
      "get": {
        "tags": ["Groups"],
        "summary": "Get all travel groups",
        "description": "Retrieve all available travel groups",
        "responses": {
          "200": {
            "description": "Groups retrieved successfully",
            "schema": {
              "type": "array",
              "items": {"$ref": "#/definitions/Group"}
            }
          }
        }
      },
      "post": {
        "tags": ["Groups"],
        "summary": "Create a travel group",
        "description": "Create a new travel group for an existing trip",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["tripId", "groupName", "gatheringPoint", "availableSpots"],
              "properties": {
                "tripId": {
                  "type": "string",
                  "example": "6659c97b3f3a123456789012"
                },
                "groupName": {
                  "type": "string",
                  "example": "Adventure Seekers"
                },
                "gatheringPoint": {
                  "type": "string",
                  "example": "Dhaka Airport Terminal 1"
                },
                "availableSpots": {
                  "type": "integer",
                  "example": 5,
                  "minimum": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "Group created successfully"},
          "400": {"description": "Bad Request - Invalid input"},
          "401": {"description": "Unauthorized"},
          "404": {"description": "Trip not found"},
          "500": {"description": "Internal Server Error"}
        }
      }
    },
    "/api/groups/{id}/join": {
      "post": {
        "tags": ["Groups"],
        "summary": "Join a travel group",
        "description": "Join an existing travel group",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Group ID"
          }
        ],
        "responses": {
          "200": {"description": "Successfully joined group"},
          "400": {"description": "Group is full or user already joined"},
          "401": {"description": "Unauthorized"},
          "404": {"description": "Group not found"}
        }
      }
    },
    "/api/groups/{id}/cancel": {
      "post": {
        "tags": ["Groups"],
        "summary": "Cancel group membership",
        "description": "Leave a travel group",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Group ID"
          }
        ],
        "responses": {
          "200": {"description": "Successfully left group"},
          "401": {"description": "Unauthorized"},
          "404": {"description": "Group not found"}
        }
      }
    },
    "/api/upload-image": {
      "post": {
        "tags": ["Media"],
        "summary": "Upload images",
        "description": "Upload multiple images to cloud storage",
        "security": [{"bearerAuth": []}],
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "images",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Image files to upload"
          }
        ],
        "responses": {
          "201": {
            "description": "Images uploaded successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {"type": "boolean"},
                "urls": {
                  "type": "array",
                  "items": {"type": "string"}
                }
              }
            }
          },
          "400": {"description": "Bad Request - No files uploaded"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
...
    "/api/photos/{id}": {
      "get": {
        "tags": ["Media"],
        "summary": "Get trip photos",
        "description": "Retrieve photos for a specific trip",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Trip ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Photos retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "photos": {
                  "type": "array",
                  "items": {"$ref": "#/definitions/Photo"}
                }
              }
            }
          },
          "401": {"description": "Unauthorized"},
          "404": {"description": "Trip not found"}
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "userID": {"type": "string"},
        "name": {"type": "string"},
        "email": {"type": "string"},
        "username": {"type": "string"},
        "bio": {"type": "string"},
        "profilePhoto": {"type": "string"},
        "createdAt": {"type": "string", "format": "date-time"}
      }
    },
    "Trip": {
      "type": "object",
      "properties": {
        "tripID": {"type": "string"},
        "userID": {"type": "string"},
        "destination": {"type": "string"},
        "duration": {"type": "integer"},
        "budget": {"type": "number"},
        "itinerary": {"type": "object"},
        "preferences": {"type": "array", "items": {"type": "string"}},
        "createdAt": {"type": "string", "format": "date-time"}
      }
    },
    "Blog": {
      "type": "object",
      "properties": {
        "blogID": {"type": "string"},
        "userID": {"type": "string"},
        "tripID": {"type": "string"},
        "title": {"type": "string"},
        "content": {"type": "string"},
        "images": {"type": "array", "items": {"type": "string"}},
        "likes": {"type": "integer"},
        "comments": {"type": "array"},
        "createdAt": {"type": "string", "format": "date-time"}
      }
    },
...
    "Photo": {
      "type": "object",
      "properties": {
        "photoID": {"type": "string"},
        "tripID": {"type": "string"},
        "url": {"type": "string"},
        "caption": {"type": "string"},
        "uploadedAt": {"type": "string", "format": "date-time"}
      }
    },
    "Pagination": {
      "type": "object",
      "properties": {
        "currentPage": {"type": "integer"},
        "totalPages": {"type": "integer"},
        "totalItems": {"type": "integer"},
        "hasNextPage": {"type": "boolean"},
        "hasPrevPage": {"type": "boolean"}
      }
    }
  }
}